% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeanSpacing.R
\name{calculateMeanSpacing}
\alias{calculateMeanSpacing}
\title{Calculate the average number of bases between adjacent SNPs.}
\usage{
calculateMeanSpacing(posData, removeRepeats = TRUE, removeZeroes = TRUE)
}
\arguments{
\item{posData}{A dataset that contains 3 columns. See \code{?snpPositions} for what this dataset must
be formatted as. Each row represents a gene, and gene has 3 attributes: its id, its length,
and the positions of all SNPs observed on it in the population.}

\item{removeRepeats}{A boolean that controls whether the function disregards repeated observations
(and in turn ignores zero-differences) while parsing. Default is TRUE.}

\item{removeZeroes}{A boolean that controls whether the function, after completing calculations,
discards averages of 0 (which indicate the gene has a single SNP). Default is TRUE.}
}
\value{
A dataframe of 2 columns:
\itemize{
    \item transcriptID - the id of the gene
    \item meanSpacing - the average number of bases between each adjacent SNP on this gene
}
}
\description{
Given a correctly formatted dataset (see \code{?snpPosition}), parse the SNP
positions for each gene and compute the average number of bases between adjacent
observed positions. If a gene contains a single SNP, then the mean spacing is
set to a default value of 0 (which can then be filtered out with \code{removeZeroes}).
The user can specify whether zero-differences (i.e. the amount of space between two
repeated observations of the same SNP) should be ignored.
}
\examples{
ms <- calculateMeanSpacing(snpPositions)
msZeroes <- calculateMeanSpacing(snpPositions, removeZeroes=FALSE)

}
