% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNPCounts.R
\name{getSNPCounts}
\alias{getSNPCounts}
\title{Provide the number of SNPs on each gene across the population.}
\usage{
getSNPCounts(posData, removeRepeats = TRUE)
}
\arguments{
\item{posData}{A dataset that contains 3 columns. See \code{?snpPositions} for what this dataset must
be formatted as. Each row represents a gene, and gene has 3 attributes: its id, its length,
and the positions of all SNPs observed on it in the population.}

\item{removeRepeats}{A boolean specifying whether observed SNPs that share the same position on the
gene are to be counted more than once. Default is \code{TRUE}.}
}
\value{
A dataframe, with each row representing a gene with the following columns:
\itemize{
    \item id - The id of the gene/transcript (unchanged from input).
    \item length - The length of the gene/transcript (unchanged from input).
    \item counts - The number of (non-)unique SNPs observed.
}
}
\description{
Given a properly formatted dataset (see \code{?snpPositions}), count the number of SNPs
found on each gene to determine the amount of population-wide variation. This function
either provides the number of unique positions on which variation occurs, or counts the
total number of observations of variation across the population.
}
\examples{
getSNPCounts(snpPositions)
getSNPCounts(snpPositions, removeRepeats=FALSE)

}
