% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeGene.R
\name{initializeGene}
\alias{initializeGene}
\title{Initialize a Gene object from the dataset.}
\usage{
initializeGene(id, posData)
}
\arguments{
\item{id}{A string denoting the id of the gene name.}

\item{posData}{A dataset that contains 3 columns. See \code{?snpPositions} for what this dataset must
be formatted as. Each row represents a gene, and gene has 3 attributes: its id, its length,
and the positions of all SNPs observed on it in the population.}
}
\value{
An S3 object of class "Gene" that summarizes a gene's degree of variation within
    the population. It contains the following fields:
\itemize{
    \item id - The name of the gene (unchanged from input)
    \item length - The length of the gene
    \item posTable - A data frame that contains a list of counts for each SNP position
    \item meanSpacing - The average number of bases between adjacent SNPs on this gene
}
}
\description{
Allows the user to identify any gene by its string ID from within a properly formatted
dataset (see \code{?snpPositions}) and initialize an object with the proper attributes
for visualization (see \code{?visualizeGene}).
}
\details{
The method returns a list (with a class defined as "Gene") with 4 fields. The \code{posTable} field
is in turn a data frame of 2 columns: the first being the SNP positions and the second being the
number of times that SNP is repeated in the population. This is used for heatmap visualization in
the final graphic.
}
\examples{
example <- "Potri.001G000800.1.v3.0"
initializeGene(example, snpPositions)
}
