% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPopulation.R
\name{plotPopulation}
\alias{plotPopulation}
\title{Provide a large-scale visualization of the variation in the population represented
by the dataset.}
\usage{
plotPopulation(posData, removeRepeats = TRUE)
}
\arguments{
\item{posData}{A dataset that contains 3 columns. See \code{?snpPositions} for what this dataset must
be formatted as. Each row represents a gene, and gene has 3 attributes: its id, its length,
and the positions of all SNPs observed on it in the population.}

\item{removeRepeats}{A boolean specifying whether observed SNPs that share the same position on the
gene are to be counted more than once. Default is \code{TRUE}.}
}
\value{
Produces a scatterplot.
}
\description{
Outputs a scatterplot using ggplot2 that plots the length of each gene against
the number of SNPs observed on that gene. By default, each locus
is counted once and only once -- repeats are ignored. However, this can be changed using
the removeRepeats parameter.
}
\examples{
plotPopulation(snpPositions)
plotPopulation(snpPositions, removeRepeats=FALSE)

}
