% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCounts.R
\name{queryCounts}
\alias{queryCounts}
\title{Query dataset for top genes}
\usage{
queryCounts(countsData, minLength = 0, minCounts = 0, sortBy = "counts")
}
\arguments{
\item{countsData}{A dataframe. Output of \code{getSNPCounts}.}

\item{minLength}{The minimum length of genes to filter by. Default is 0 (removes nothing).}

\item{minCounts}{The minimum number of observed SNPs to filter by. Default is 0 (removes nothing).}

\item{sortBy}{A string that decides the sorting setting of the output. If none of the below arguments
    are used, use default sorting. Potentional arguments:
\itemize{
    \item \code{counts} - The default argument. Sort by decreasing SNP counts.
    \item \code{length} - Sort by decreasing gene length
    \item \code{none} - No sorting is performed, and the subset is returned as is.
}}
}
\value{
A subset of the countsData data frame that satisfies the user inputs.
}
\description{
Filter and sort the given dataset (must be of counts, output by \code{getSNPCounts}) based on user-
defined parameters.
}
\examples{
countsData <- getSNPCounts(snpPositions)
topHits1 <- queryCounts(countsData, 10000, 150)
topHits2 <- queryCounts(countsData, 10000, 150, sortBy="length")
topHits3 <- queryCounts(countsData, 10000, 150, sortBy="asdf")

}
