% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runViewer.R
\name{runViewer}
\alias{runViewer}
\title{Run associated Shiny webpage application}
\usage{
runViewer()
}
\description{
Provides the user with a visual interface with which to explore the population data
and display any genes of interest.
}
\examples{
runViewer()
}
\references{
Winston Chang, Joe Cheng, JJ Allaire, Yihui Xie and Jonathan McPherson (2020). shiny: Web
    Application Framework for R. R package version 1.5.0. https://CRAN.R-project.org/package=shiny
}
