% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snpPositions}
\alias{snpPositions}
\title{SNP Position data from a GWAS experiment.}
\format{
A matrix with columns:
\describe{
 \item{transcript_id}{The name of the gene/transcript}
 \item{gene_length}{The length of the gene cDNA, in bases}
 \item{snp_position_list}{The positions of all SNPs found on this gene
 across the population, separated by bars '|'.}
}
}
\source{
???
}
\usage{
snpPositions
}
\description{
Where does the data come from?
}
\examples{
\dontrun{snpPositions}

}
\keyword{datasets}
