% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeGene.R
\name{visualizeGene}
\alias{visualizeGene}
\title{Provide a graphical visualization of a given gene and its degree of variation.}
\usage{
visualizeGene(gene)
}
\arguments{
\item{gene}{An S3 object of class Gene. The gene to be visualized. See \code{?initializeGene} for
information on this class' attributes.}
}
\value{
Produces an R graphic visualization of the given gene.
}
\description{
Draws a gray rectangle on an R plot environment to represent the full length of the gene.
x-axis labels are the position numbers of the gene's cDNA. On top of the rectangle,
vertical coloured line segments represent positions on which SNPs are found -- the degree of
redness is a heatmap colour mapping that depicts its relative frequency.
}
\examples{
gene <- initializeGene("Potri.001G345300.1.v3.0", snpPositions)
visualizeGene(gene)

}
