% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeanSpacing.R
\name{calculateMeanSpacing}
\alias{calculateMeanSpacing}
\title{Calculate the average number of bases between adjacent SNPs.}
\usage{
calculateMeanSpacing(posData, removeRepeats = TRUE)
}
\arguments{
\item{posData}{A dataset that contains 3 columns. See \code{?snpPositions} for what this dataset must
be formatted as. Each row represents a gene, and gene has 3 attributes: its id, its length,
and the positions of all SNPs observed on it in the population.}

\item{removeRepeats}{A boolean that controls whether the function disregards repeated observations
(and in turn ignores zero-differences) while parsing. Default is TRUE.}
}
\value{
A dataframe of 2 columns:
\itemize{
    \item transcriptID - the id of the gene
    \item meanSpacing - the average number of bases between each adjacent SNP on this gene
}
}
\description{
Given a correctly formatted dataset (see \code{?snpPositions}), parse the SNP
positions for each gene and compute the average number of bases between adjacent
observed positions. Calculations are padded by inserting fictitious SNPs at the
beginning and end of a gene (indices 0 and at length + 1). This is to demonstrate the
sparsity of SNPs on genes with few observed loci. The user can specify whether
zero-differences (i.e. the amount of space between two repeated observations of
the same SNP) should be ignored.
}
\examples{
ms <- calculateMeanSpacing(snpPositions)

}
