% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snpPositions}
\alias{snpPositions}
\title{Non-synonymous coding region SNP Position Data from Next Generation Sequencing}
\format{
A matrix with columns:
\describe{
 \item{transcript_id}{The name of the gene/transcript}
 \item{gene_length}{The length of the gene cDNA, in bases}
 \item{snp_position_list}{The positions of all SNPs found on this gene
 across the population, separated by bars '|'. They are sorted in increasing
 order, and repeated indices represent the same SNP position being observed in
 more than one individual/genotype.}
}
}
\source{
Provart Lab, Department of Cells and Systems Biology, University of Toronto.
}
\usage{
snpPositions
}
\description{
Genetic variation data obtained by parsing genomic data (initially stored in Variant Call Format) from
    RNA-Seq experiments on Poplar individuals.
}
\examples{
\dontrun{snpPositions}

}
\keyword{datasets}
